+++ docker/changelog.md
@@ -5,6 +5,17 @@ All notable changes to this component will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0
+
+### Added
+
+* option to include the namespace in the csv reports
+* `LABEL`s in docker image
+
+### Changed
+
+* updated base image: `bitnami/kubectl:1.30` -> `bitnami/kubectl:1.33`
+
 ## 1.0.1
 
 ### Fixed